# Conflicts:
#	features/logistic/localizationchooseaddress/src/main/java/com/tokopedia/localizationchooseaddress/ui/bottomsheet/ChooseAddressBottomSheet.kt
#	features/logistic/localizationchooseaddress/src/main/java/com/tokopedia/localizationchooseaddress/ui/widget/ChooseAddressWidget.kt
#	libraries/applink/src/test/java/tokopedia/applink/deeplink/DeepLinkMapperCustomerAppTest.kt

+++ features/merchant/gold_merchant_common/src/main/java/com/tokopedia/gm/common/presentation/model/ShopInfoPeriodUiModel.kt
@@ -4,7 +4,7 @@ package com.tokopedia.gm.common.presentation.model
 data class ShopInfoPeriodUiModel(var isNewSeller: Boolean = false,
                                  var isEndTenureNewSeller: Boolean = false,
                                  var periodType: String = "",
-                                 var shopAge: Int = 0,
+                                 var shopAge: Long = 0,
                                  var periodStartDate: String = "",
                                  var periodEndDate: String = ""
 )
